DCC-541 - Fixes a bunch of things, restyle login, handle 404s better
Former-commit-id: 2639bbd497d63e9161d652345f76ed611d942e59
Former-commit-id: 5d3edce5f27d1533ce0b47ab088a650d608eeb31

+++ client/app/controllers/auth_controller.coffee
@@ -28,13 +28,17 @@ module.exports = class AuthController extends Controller
   historyURL: 'auth'
 
   logout: ->
-    # Save the current location so you don't lose your
-    # place on logout
-    location = window.location
-    localStorage.clear()
-    Chaplin.mediator.publish '!logout'
-    # I would rather use a Chaplin or Backbone method
-    # here, but this is the only way I can get the url
-    # to change without a refresh
-    window.location = location
-      
+    if Chaplin.mediator.user
+
+      # Save the current location so you don't lose your
+      # place on logout
+      location = window.location
+      localStorage.clear()
+      Chaplin.mediator.publish '!logout'
+      # I would rather use a Chaplin or Backbone method
+      # here, but this is the only way I can get the url
+      # to change without a refresh
+      window.location = location
+
+    else
+      Chaplin.mediator.publish '!startupController', 'release', 'list'
